make the iasync dependency on gvision conditional
